+++ xulixoatatca.php
@@ -1,25 +1,30 @@
 <?php 
+  session_start();
+  $username = $_SESSION['username'];
+  if (isset($username)) {
+    $link = mysqli_connect("localhost", "root", "root", "dulieu");
+    if(!$link) die(" Connection fail" .mysqli_connect_error());
+    if (isset($_POST['xoa'])) {
+        
+      $deleteId = $_POST['xoa'];
+    
+      // implode function break the array in to string 
+      $deleteId = implode("','", $deleteId);
 
-$link = mysqli_connect("localhost", "root", "root", "dulieu");
-if(!$link) die(" Connection fail" .mysqli_connect_error());
-if (isset($_POST['xoa'])) {
-		
-  $deleteId = $_POST['xoa'];
- 
-  // implode function break the array in to string 
-  $deleteId = implode("','", $deleteId);
+      // delete all nhan vien co trong cac phong ban nay
+      $queryDeleteNhanvien  = "DELETE FROM nhanvien WHERE idpb IN ('$deleteId')";
+      $result = mysqli_query($link, $queryDeleteNhanvien);
+      
 
-  // delete all nhan vien co trong cac phong ban nay
-  $queryDeleteNhanvien  = "DELETE FROM nhanvien WHERE idpb IN ('$deleteId')";
-  $result = mysqli_query($link, $queryDeleteNhanvien);
-   
+      $query  = "DELETE FROM phongban WHERE idpb IN ('$deleteId')";
 
-  $query  = "DELETE FROM phongban WHERE idpb IN ('$deleteId')";
-
-  $result = mysqli_query($link, $query);
-  
-}
-mysqli_close($link);
-header("Location:/dulieu/phongban.php");
+      $result = mysqli_query($link, $query);
+      
+    }
+    mysqli_close($link);
+    header("Location:/dulieu/phongban.php");
+  } else {
+    header("Location:/dulieu/login.php");
+  }
 
 ?>
